+++ ohrms_loan/models/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
 from . import hr_loan
 from . import hr_payroll
+from . import res_config_settings