+++ ohrms_loan/doc/RELEASE_NOTES.md
@@ -11,9 +11,16 @@
 - Fixed TypeError Git Issue #13
 
 #### 12/12/2018
-#### Version 12.0.1.0.1
+#### Version 12.0.2.0.0
 ##### CHANGE
 - Exception class changed
 - State visibility changed
 - Restriction added for deletion
+
+#### 12/12/2018
+#### Version 12.0.3.0.0
+##### ADD
+- Added configuration settings for multiple loan eligibility with boolean toggle.
+- Implemented three conditional checks: by Years of Service, and by Job Position.
+- Resolved loan eligibility validation issue where pending installments blocked new loan requests unnecessarily.